requirements.txt
Flask==3.0.3
PyYAML==6.0.1
python-dotenv==1.0.1
openai==1.35.8
